[fix] operator input in chat
